+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+### Features
+
 - SentryFeedbackWidget Improvements ([#2964](https://github.com/getsentry/sentry-dart/pull/2964))
  - Pick screenshots from camera roll for feedback
  - Capture a device screenshot for feedback
@@ -10,6 +12,12 @@
 ```dart
 options.feedback.showBranding = false;
 ```
+
+## 9.0.0-RC.4
+
+### Enhancements
+
+- Replay: Mask RichText Widgets ([#2975](https://github.com/getsentry/sentry-dart/pull/2975))
 
 ## 9.0.0-RC.3